// bo-front/src/app/profiles/pages/ambassador/pages/tree/pages/history-range/commons/components/my-rewards/commons/interfaces/partnerList.ts
 export class BonoCarListResponseDTO {
  color: string;
  stateName: string;
  sponsorLevel: number;
  sponsorName: string;
  installmentNumber: string;
  paymentDate: string;
  financedPaymentUSD: number;
  interestUSD: number;
  amortizationUSD: number;
  splitInitialPaymentUSD: number;
  rankBonusUSD: number;
  finalPaymentUSD: number;
  remainingBalanceUSD: number;

}
 